+++ src/constants/keyboard.const.ts
@@ -1,6 +1,7 @@
 export const KEY_ENTER = 'Enter';
 export const KEY_ESCAPE = 'Escape';
 export const KEY_TAB = 'Tab';
+export const KEY_F2 = 'F2';
 export const KEY_ARROW_UP = 'ArrowUp';
 export const KEY_ARROW_DOWN = 'ArrowDown';
 export const KEY_ARROW_LEFT = 'ArrowLeft';
@@ -9,4 +10,6 @@ export const KEY_HOME = 'Home';
 export const KEY_END = 'End';
 export const KEY_PAGE_UP = 'PageUp';
 export const KEY_PAGE_DOWN = 'PageDown';
+export const KEY_C = 'c';
+export const KEY_V = 'v';
 export const BOOLEAN_TRUE_VALUE = 'true';